backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
pydantic==2.12.4
starlette==0.37.2
anyio==4.11.0
sniffio==1.3.1
idna==3.11
email-validator==2.3.0
elevenlabs==2.16.0
google-generativeai>=0.8.0
